Moved the CryptoError into error.rs with all the other errors
